worker/requirements.txt
mlx-audio>=0.2.1
pyannote.audio>=4.0.0
torch>=2.5.0
soundfile>=0.12.1
numpy>=1.26.0
